It works, kinda sorta. Vector->std::vector, various multiplatform fixes.

+++ src/Simfile/Tempo.cpp
@@ -37,7 +37,8 @@ void Tempo::sanitize(const Chart* owner) {
     auto bpmc = segments->begin<BpmChange>();
     auto end = segments->end<BpmChange>();
     if (bpmc == end || bpmc->row != 0) {
-        double bpm = (bpmc != end) ? bpmc->bpm : SIM_DEFAULT_BPM;
+        double bpm =
+            (bpmc != end) ? bpmc->bpm : static_cast<double>(SIM_DEFAULT_BPM);
         segments->insert(BpmChange(0, bpm));
         std::string suffix;
         if (owner) {
